Migrate D455_frame_drops test from legacy to pytest
